UPDATE Terraform scripts, IAM policy files, Lambda functions and documentation to deploy PDS Nucleus in MCP Test and Prod environments.
Refer to the issue:
https://github.com/NASA-PDS/nucleus/issues/89
