+++ .secrets.baseline
@@ -2,6 +2,9 @@
   "version": "1.4.0",
   "plugins_used": [
     {
+      "name": "AbsolutePathDetectorExperimental"
+    },
+    {
       "name": "ArtifactoryDetector"
     },
     {
@@ -85,6 +88,10 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
+      "path": "detect_secrets.filters.common.is_baseline_file",
+      "filename": ".secrets.baseline"
+    },
+    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -119,16 +126,9 @@
       "path": "detect_secrets.filters.regex.should_exclude_file",
       "pattern": [
         "\\.secrets..*",
-        "\\.pre-commit-config\\.yaml",
         "\\.git.*",
-        "\\.mypy_cache",
-        "\\.pytest_cache",
-        "\\.tox",
-        "\\.venv",
-        "venv",
-        "dist",
-        "build",
-        ".*\\.egg-info"
+        "\\.pre-commit-config\\.yaml",
+        "target"
       ]
     }
   ],
@@ -139,7 +139,8 @@
         "filename": "setup.cfg",
         "hashed_secret": "3a6d7aa49a8e4a2fe32a5cd0e53da9cb96bd8d29",
         "is_verified": false,
-        "line_number": 22
+        "line_number": 22,
+        "is_secret": false
       }
     ],
     "terraform/README.md": [
@@ -148,23 +149,26 @@
         "filename": "terraform/README.md",
         "hashed_secret": "f2d4e04179e44fa7386b985ac3c7ee4d95dfd65d",
         "is_verified": false,
-        "line_number": 83
+        "line_number": 87,
+        "is_secret": false
       },
       {
         "type": "AWS Sensitive Information (Experimental Plugin)",
         "filename": "terraform/README.md",
-        "hashed_secret": "058af5f48af6e6e280495dcf5b889f92dd36863d",
+        "hashed_secret": "659a4d010b74afeddbcb9c4e8eae01f4390eeacc",
         "is_verified": false,
-        "line_number": 84
+        "line_number": 88,
+        "is_secret": false
       },
       {
         "type": "AWS Sensitive Information (Experimental Plugin)",
         "filename": "terraform/README.md",
-        "hashed_secret": "431aad08e6c7e9babc8c125c7504e5ba71dfe7a7",
+        "hashed_secret": "bd3b85b91cb8cf6cfc6a4adc7a2505714939505b",
         "is_verified": false,
-        "line_number": 84
+        "line_number": 88,
+        "is_secret": false
       }
     ]
   },
-  "generated_at": "2023-11-17T20:05:20Z"
+  "generated_at": "2024-05-10T00:20:21Z"
 }
