+++ terraform/providers.tf
@@ -10,4 +10,10 @@ terraform {
 
 provider "aws" {
   region = var.region
+
+  default_tags {
+    tags = {
+      product     = "PDS Nucleus"
+    }
+  }
 }
