+++ terraform/terraform-modules/ecs-ecr/docker/s3-to-efs-copy/entrypoint.sh
@@ -1,0 +1,86 @@
+#!/bin/bash
+# Copyright 2024, California Institute of Technology ("Caltech").
+# U.S. Government sponsorship acknowledged.
+#
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice,
+# this list of conditions and the following disclaimer.
+# * Redistributions must reproduce the above copyright notice, this list of
+# conditions and the following disclaimer in the documentation and/or other
+# materials provided with the distribution.
+# * Neither the name of Caltech nor its operating division, the Jet Propulsion
+# Laboratory, nor the names of its contributors may be used to endorse or
+# promote products derived from this software without specific prior written
+# permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+
+# Check the user or role that made the call
+aws sts get-caller-identity
+
+# Read values from command line arguments
+
+EFS_CONFIG_DIR=$1
+OPERATION=$2
+
+if [ "$OPERATION" = "DELETE" ]
+then
+
+  filename="$EFS_CONFIG_DIR"/files_created.txt
+  echo "Removing data files listed in $filename"
+
+  while read -r line; do
+      file_to_delete="$line"
+      echo "Deleting- $file_to_delete"
+      rm "$file_to_delete"
+  done < "$filename"
+
+else # Not a delete command
+
+  echo "Copying data files in $EFS_CONFIG_DIR/data_file_list.txt"
+  filename=$EFS_CONFIG_DIR/data_file_list.txt
+  while read -r line; do
+      s3_url_of_file="$line"
+      echo "Name read from file - $s3_url_of_file"
+      file_path=$(awk '{print substr($0, 6)}' <<< "$s3_url_of_file")
+      dir_path=$(dirname "$file_path")
+      str_to_replace="s3://"
+      replace_with="/mnt/data/"
+      target_location="${s3_url_of_file//$str_to_replace/$replace_with}"
+      aws s3 cp "$s3_url_of_file" "$target_location"
+      echo "$target_location" >> "$EFS_CONFIG_DIR"/files_created.txt
+
+      # Extract .fz files
+      if [[ $target_location == *.fz ]]
+      then
+        echo "Extracting $target_location..."
+        extracted_file_name=${target_location%.*}
+
+        if [ -f "$extracted_file_name" ] ; then
+            rm "$extracted_file_name"
+        fi
+
+        funpack -v "$target_location"
+        echo "$extracted_file_name" >> "$EFS_CONFIG_DIR"/files_created.txt
+      fi
+
+  done < "$filename"
+
+fi
+
+
