+++ terraform/terraform-modules/ecs-ecr/docker/s3-to-efs-copy/s3_files_list_to_copy.txt
@@ -1,0 +1,9 @@
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWF.LBL
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWF.TAB
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWF.xml
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWG.LBL
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWG.TAB
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWG.xml
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWI.LBL
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWI.TAB
+s3://pds-nucleus-staging/messenger-data/MSGRMDS_6001/CALIB/SUPPORT/MDISBPWI.xml
