+++ terraform/terraform-modules/ecs-ecr/ecs_task_execution_role_iam_policy.json
@@ -1,0 +1,20 @@
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Effect": "Allow",
+      "Action": [
+        "ecr:GetAuthorizationToken",
+        "ecr:BatchCheckLayerAvailability",
+        "ecr:GetDownloadUrlForLayer",
+        "ecr:BatchGetImage",
+        "logs:CreateLogStream",
+        "logs:PutLogEvents"
+      ],
+      "Resource": [
+        "arn:aws:ecr:*:<AWS-ACCOUNT-ID>:*",
+        "arn:aws:logs:*:<AWS-ACCOUNT-ID>:*"
+      ]
+    }
+  ]
+}
