+++ terraform/terraform-modules/ecs-ecr/ecs_task_role_iam_policy.json
@@ -1,0 +1,33 @@
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Effect": "Allow",
+      "Action": [
+        "logs:CreateLogStream",
+        "ecr:GetDownloadUrlForLayer",
+        "ecr:BatchGetImage",
+        "ecr:GetAuthorizationToken",
+        "logs:PutLogEvents",
+        "ecr:BatchCheckLayerAvailability",
+        "s3:Get*",
+        "s3:List*",
+        "s3:Describe*",
+        "s3-object-lambda:Get*",
+        "s3-object-lambda:List*",
+        "es:ESHttpGet",
+        "elasticfilesystem:ClientMount",
+        "elasticfilesystem:ClientRootAccess",
+        "elasticfilesystem:ClientWrite",
+        "elasticfilesystem:DescribeMountTargets"
+      ],
+      "Resource": [
+        "arn:aws:logs:*:<AWS-ACCOUNT-ID>:*",
+        "arn:aws:ecr:*:<AWS-ACCOUNT-ID>:*",
+        "arn:aws:s3:*:<AWS-ACCOUNT-ID>:*",
+        "arn:aws:es:*:<AWS-ACCOUNT-ID>:*",
+        "arn:aws:elasticfilesystem:*:<AWS-ACCOUNT-ID>:*"
+      ]
+    }
+  ]
+}
