+++ terraform/terraform-modules/efs/efs.tf
@@ -4,8 +4,14 @@ resource "aws_efs_file_system" "nucleus_efs" {
   creation_token = "nucleus_efs_token"
 
   tags = {
-    Name = "Nucleus"
+    Name = "Nucleus EFS"
   }
+}
+
+resource "aws_efs_mount_target" "pds_nucleus_efs_mount_target" {
+  file_system_id  = aws_efs_file_system.nucleus_efs.id
+  subnet_id       = var.subnet_ids[0]
+  security_groups = [var.nucleus_security_group_id]
 }
 
 resource "aws_efs_access_point" "root" {
@@ -21,41 +27,36 @@ resource "aws_efs_access_point" "root" {
   }
 }
 
-resource "aws_efs_access_point" "scripts" {
+
+resource "aws_efs_access_point" "pds-data" {
 
   file_system_id = aws_efs_file_system.nucleus_efs.id
 
   root_directory {
-    path = "/registry/docker/scripts"
+    path = "/pds-data"
+
+    creation_info {
+      owner_gid   = 1000
+      owner_uid   = 1000
+      permissions = 0755
+    }
   }
 
+
   tags = {
-    Name = "scripts access point"
+    Name = "PDS Data access point"
   }
 }
 
-resource "aws_efs_access_point" "registry-loader-waits-for-elasticsearch" {
 
-  file_system_id = aws_efs_file_system.nucleus_efs.id
-
-  root_directory {
-    path = "/registry/docker/scripts/registry-loader-waits-for-elasticsearch.sh"
-  }
-
-  tags = {
-    Name = "registry-loader-waits-for-elasticsearch access point"
-  }
+output "efs_file_system_id" {
+  value = aws_efs_file_system.nucleus_efs.id
 }
 
-resource "aws_efs_access_point" "default-config" {
+output "efs_access_point_id_root" {
+  value = aws_efs_access_point.root.id
+}
 
-  file_system_id = aws_efs_file_system.nucleus_efs.id
-
-  root_directory {
-    path = "/registry/docker/default-config"
-  }
-
-  tags = {
-    Name = "default-config access point"
-  }
+output "efs_access_point_id_pds-data" {
+  value = aws_efs_access_point.pds-data.id
 }
