+++ terraform/terraform-modules/mwaa-env/mwaa_iam_policy.json
@@ -1,0 +1,75 @@
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Effect": "Allow",
+      "Action": "airflow:PublishMetrics",
+      "Resource": "arn:aws:airflow:*:<AWS-ACCOUNT-ID>:*"
+    },
+    {
+      "Effect": "Allow",
+      "Action": [
+        "s3:GetObject*",
+        "s3:GetBucket*",
+        "s3:ListAllMyBuckets",
+        "s3:List*"
+      ],
+      "Resource": [
+        "arn:aws:s3:*:<AWS-ACCOUNT-ID>:*"
+      ]
+    },
+    {
+      "Effect": "Allow",
+      "Action": [
+        "ecs:RunTask",
+        "ecs:DescribeTasks"
+      ],
+      "Resource": [
+        "arn:aws:ecs:*:<AWS-ACCOUNT-ID>:*"
+      ]
+    },
+    {
+      "Effect": "Allow",
+      "Action": [
+        "logs:CreateLogStream",
+        "logs:CreateLogGroup",
+        "logs:PutLogEvents",
+        "logs:GetLogEvents",
+        "logs:GetLogRecord",
+        "logs:GetLogGroupFields",
+        "logs:GetQueryResults",
+        "logs:DescribeLogGroups"
+      ],
+      "Resource": [
+        "arn:aws:logs:*:<AWS-ACCOUNT-ID>:*"
+      ]
+    },
+    {
+      "Effect": "Allow",
+      "Action": "cloudwatch:PutMetricData",
+      "Resource": "arn:aws:cloudwatch:*:<AWS-ACCOUNT-ID>:*"
+    },
+    {
+      "Effect": "Allow",
+      "Action": "iam:PassRole",
+      "Resource": "*"
+    },
+    {
+      "Effect": "Allow",
+      "Action": [
+        "kms:Decrypt",
+        "kms:DescribeKey",
+        "kms:GenerateDataKey*",
+        "kms:Encrypt"
+      ],
+      "Resource": "arn:aws:kms:*:<AWS-ACCOUNT-ID>:*",
+      "Condition": {
+        "StringLike": {
+          "kms:ViaService": [
+            "sqs.us-west-2.amazonaws.com"
+          ]
+        }
+      }
+    }
+  ]
+}
